Refactor audio package into module-based directory structure (#2)
