+++ packages/audio/src/index.mjs
@@ -1,6 +1,6 @@
-export { Synthesizer } from './Synthesizer.mjs';
-export { Wave } from './Wave.mjs';
-export { presetWaveStrategy } from './waveStrategies/presetWaveStrategy.mjs';
-export { tableWaveStrategy } from './waveStrategies/tableWaveStrategy.mjs';
-export { noiseWaveStrategy } from './waveStrategies/noiseWaveStrategy.mjs';
-export { lfsr } from './utils/lfsr.mjs';
+export { Synthesizer } from './synthesizer/Synthesizer.mjs';
+export { Wave } from './synthesizer/Wave.mjs';
+export { presetWaveStrategy } from './synthesizer/waveStrategies/presetWaveStrategy.mjs';
+export { tableWaveStrategy } from './synthesizer/waveStrategies/tableWaveStrategy.mjs';
+export { noiseWaveStrategy } from './synthesizer/waveStrategies/noiseWaveStrategy.mjs';
+export { lfsr } from './synthesizer/utils/lfsr.mjs';
