+++ packages/audio/src/synthesizer/Synthesizer.mjs
