+++ packages/audio/src/synthesizer/Wave.mjs
