+++ packages/audio/src/synthesizer/utils/lfsr.mjs
