+++ packages/audio/src/synthesizer/utils/spn2freq.mjs
