+++ packages/audio/src/synthesizer/waveStrategies/noiseWaveStrategy.mjs
