+++ packages/audio/src/synthesizer/waveStrategies/presetWaveStrategy.mjs
