+++ packages/audio/src/synthesizer/waveStrategies/tableWaveStrategy.mjs
