+++ packages/audio/test/synthesizer/waveStrategies/presetWaveStrategy.test.js
@@ -1,11 +1,11 @@
 import { test, mock } from 'node:test';
 import assert from 'node:assert/strict';
-import { presetWaveStrategy } from '../../src/waveStrategies/presetWaveStrategy.mjs';
+import { presetWaveStrategy } from '../../../src/synthesizer/waveStrategies/presetWaveStrategy.mjs';
 
 global.__DEV__ = true;
 
 // Stub: Frequency of 'A5' (Self-made utility func)
-import * as spn from '../../src/utils/spn2freq.mjs';
+import * as spn from '../../../src/synthesizer/utils/spn2freq.mjs';
 mock.fn(spn, 'spn2freq', () => 880);
 
 // Mock: OscillatorNode for testing (Web Audio API is not available in Node.js)
